+++ ssh_entity/src/main/java/com/ywx/erp/entity/StoredetailDo.java
@@ -11,26 +11,26 @@ public class StoredetailDo implements Serializable {
 	public static final String STATE_NOT_OUT = "0";	//未出库
 	public static final String STATE_OUT = "1";		//已出库
 
-	private Long uuid;			//编号
-	private Long storeuuid;		//仓库编号
+	private Integer uuid;			//编号
+	private Integer storeuuid;		//仓库编号
 	private String storename;	//仓库名称
-	private Long goodsuuid;		//商品编号
+	private Integer goodsuuid;		//商品编号
 	private String goodsname;	//商品名称
-	private Long num;			//数量
+	private Integer num;			//数量
 
-	public Long getUuid() {
+	public Integer getUuid() {
 		return uuid;
 	}
 
-	public void setUuid(Long uuid) {
+	public void setUuid(Integer uuid) {
 		this.uuid = uuid;
 	}
 
-	public Long getStoreuuid() {
+	public Integer getStoreuuid() {
 		return storeuuid;
 	}
 
-	public void setStoreuuid(Long storeuuid) {
+	public void setStoreuuid(Integer storeuuid) {
 		this.storeuuid = storeuuid;
 	}
 
@@ -42,11 +42,11 @@ public class StoredetailDo implements Serializable {
 		this.storename = storename;
 	}
 
-	public Long getGoodsuuid() {
+	public Integer getGoodsuuid() {
 		return goodsuuid;
 	}
 
-	public void setGoodsuuid(Long goodsuuid) {
+	public void setGoodsuuid(Integer goodsuuid) {
 		this.goodsuuid = goodsuuid;
 	}
 
@@ -58,11 +58,11 @@ public class StoredetailDo implements Serializable {
 		this.goodsname = goodsname;
 	}
 
-	public Long getNum() {
+	public Integer getNum() {
 		return num;
 	}
 
-	public void setNum(Long num) {
+	public void setNum(Integer num) {
 		this.num = num;
 	}
