+++ ssh_service/src/main/java/com/ywx/erp/service/OrderdetailService.java
@@ -4,7 +4,7 @@ import com.ywx.erp.entity.OrderdetailDo;
 
 public interface OrderdetailService extends BaseService<OrderdetailDo> {
 
-    void doInStore(long uuid, long storeId, long empId);
+    void doInStore(int uuid, int storeId, int empId);
 
-    void doOutStore(long uuid, long storeId, long empId);
+    void doOutStore(int uuid, int storeId, int empId);
 }
